Re-coded how to handle "default_password"
Variable |string BaseParse::default_password| was treated in other way than its original design. This commit tries to correct it.

Initially, I created a variable "default_password" and used it as the default  password if no option for password is given. It itself contained only trivial messages. So if all options were parsed but the password was found its default (the same as default_password), no password had been given and a method for interactively taking password was invoked. At that time, only a "default_password". No "default_account". And the code was also scripted this way.

On Oct. and Nov., I forgot this original design: having found that there was only not a "default_account" for "default_password" that the connection could not be established, I added a "default_account" hoping to use this pair of parameters to connect the database if either was omitted in the terminal arguments. Recently, this misunderstanding was finally discovered and this new branch emerges.

Details available in "change.log"